refactor: move MotionCommand to types/types.py
